feat: separate state status with indicator in status subcommand
To make changing the indicator more flexible, i feel like we need to
separate the state status with the indicator.

So the plan is to have state status consistent with value such as
`focus` and `break`, and then there's a function
`DetermineStateIndicator()` that the user can change for state
indicator when using `status` subcommand.