fix: prevent changing state using current instead of subcommand
Previously the warning message is using the subcommand state instead of
the current state.

For example, let's say that previously we activate the focus session
and then we want to trigger the break session. Previous mechanism
will show us warning that the `break` session is on going instead of
the `focus` session, even though the current session is the focus
session.