feat: add reset subcommand
This subcommand is to reset the current state. Reset in here means
deleting the status file. So if there's an error, that's most likely
related to removing file error.

Why didn't i just change the end time? Because deleting file is more
convenient if i want to change the format of the status file.