fix(utils): condition check status file never nil
Because `json.Unmarshal()` can not take `nil` pointer, we need to
provide the pointer with default value. Which means that the return
value of `ReadStatusFile()` is never `nil`.

The effect of that is the conditional `if status != nil` will always
turn true, which is not expected. So let's change the condition to
check if the file exist by using `State != ""`.

If the file not exist, status state will be empty string. And if the
file is exist, status state should not be empty string. Let's see how
it goes.