fix: is_notify status changed to false when pausing the timer
The problem is that we only declare status state, so the other status
info is using default value. And because `is_notify` is a boolean type,
it use `false` as default value. That's why `is_notify` status always
change to `false` even though the previous info is `true`.

To solve that, just read the entire status info and change the required
info on the go.

Maybe i can simpify this into using `ReadStatusFile()` instead? We'll
see.