test: add case for subcommand (#2)
* style(utils-test): use capitalization on `got ..., want ...`

The format of `t.Errorf()` is something like `file.go:69: message`, so
if we use lower case for the message, we might miss the error message.

Let's try using capitalization at the beginning of the message and see
how it goes.

* feat(utils): add test setup temporary file and create status file

* test: add case for pause subcommand

* test: add is_notify correct case for pause subcommand

* chore: rename file internal/on-going.go to internal/on_going.go

Make the filename scheme consistent using underscore instead of dash.

* refactor: put directory path as TestSetupTempFile() parameter

This is so that we can debug the generated json file in test by
replacing the temporary directory with `/tmp`.

* refactor(utils): separate cleanup file test into another function

This is so that i can use it separately from TestSetupStatusFile().
Sometimes i need to test the scenario where the status file is not
exist yet and need to reset the file state so that we can reuse the
file.

* refactor(pause_test): move cleanup function before looping cases

The cleanup function is called in "last added, first called" order, so
let's put it before looping cases to make sure that it's called last.

* Revert "refactor(utils): separate cleanup file test into another function"

This reverts commit 2df4d632eb1cb937a4d2826bd7d3c30a7465a164.

* test: add case for end time in on going function

Test whether the end time result correct or not.

* test: add case for is_notify i on going function

Test whether `is_notify` is true or false depending on input use notify
and notify command exist.

* chore: remove todo in reset subcommand

Feels like we don't need to create test for reset functionality as it's
basically only deleting the status file.

* test: add case for status pause indicator

Test the status indicator when we have pause point in status file.

* test: add case for status remaining indicator

* ci: add test pipeline on pull request

* ci: add synchronize type event in test pipeline

To make sure that the pipeline get triggered in new commit (?). Let's
see how it goes.

* test: fix TestOnGoingEndTime first cases

Let's see if this changes can fix the failed test.

* test: use `Compare()` instead of directly compare time value

Let's see if using `Compare()` actually solve the failing test on
github CI.

* chore: add comment about cleanup in TestOnGoingEndTime/not_exist